squash utilityTypes and types and separate with a helpful comment
