+++ packages/core/src/core/procedureBuilder.ts
@@ -1,7 +1,11 @@
 import { getTRPCErrorFromUnknown, TRPCError } from '../error/TRPCError';
-import type { GetRawInputFn } from '../types';
-import { mergeWithoutOverrides } from '../utilityFunctions';
-import type { MaybePromise, Overwrite, Simplify } from '../utilityTypes';
+import type {
+  GetRawInputFn,
+  MaybePromise,
+  Overwrite,
+  Simplify,
+} from '../types';
+import { mergeWithoutOverrides } from '../utils';
 import type {
   AnyMiddlewareFunction,
   MiddlewareBuilder,
